add read history api
