+++ src/electron/main.js
@@ -29,7 +29,7 @@ const {
 const {validateJsonFile} = require('./utils/validateFile')
 const { standardResultCodes, decryptResultCodes } = require('./constants')
 const { generateSignedExitMessage } = require('./utils/Eth2Deposit')
-
+const { getHistoryRecordsByPage } = require("./utils/historyUtils")
 
 function createWindow() {
     // Create a new window
@@ -338,5 +338,16 @@ ipcMain.on("req-get-password", async (event, args) => {
     } catch (error) {
         logger.error("Error getting password", error)
         event.sender.send("receive-get-password", standardResultCodes.ERROR, '' , error.message)
+    }
+})
+
+ipcMain.on("req-history-page", async (event, args) => {
+    const [page] = args
+    try {
+        const historyRecords = await getHistoryRecordsByPage(page);
+        event.sender.send("receive-history-page", standardResultCodes.SUCCESS, historyRecords, '')
+    } catch (error) {
+        logger.error("Error getting password", error)
+        event.sender.send("receive-history-page", standardResultCodes.ERROR, '' , error.message)
     }
 })
