+++ src/react/components/DBExplorerComponents/DBExplorer.tsx
@@ -1,29 +1,6 @@
-import { useState, useEffect } from "react";
-import { useFormContext } from "react-hook-form";
-import raisedWidgetStyle from "../../styleClasses/widgetBoxStyle";
-import {
-  Center,
-  Box,
-  Text,
-  Tabs,
-  Tab,
-  TabList,
-  TabPanels,
-  TabPanel
-} from "@chakra-ui/react";
-import AddressSelect from "./AddressSelect";
-import DataTable from "./DataTable";
-import useGetPassword from "../../hooks/useGetPassword";
-
-interface StakerMap {
-  [stakerAddress: string]: StakerInfo
-}
-
-interface StakerInfo {
-  mnemonicCount: number
-  mnemonics: {[idx: string]: string}
-  validators: {[idx: string]: string}
-}
+import { ScaleFade } from '@chakra-ui/react'
+import HistoryWidget from './HistoryWidget'
+import SavedDataWidget from './SavedDataWidget'
 
 interface DBExplorerProps {
   password: string
@@ -31,87 +8,33 @@ interface DBExplorerProps {
   selectedOption: number
 }
 
-const DBExplorer = (props: DBExplorerProps) => {
-  const [allStakers, setAllStakers] = useState<StakerMap>({});
-  const [currAddress, setCurrAddress] = useState("");
-  const [addressList, setAddressList] = useState([]);
-  const [selectedTab, setSelectedTab] = useState<number>(0);
+const DBDropDownOptions = {
+  "Saved Mnemonics & Validator keys": 0,
+  "Staker History": 1
+}
 
-  const {watch} = useFormContext()
-  const dbPassword = watch("loginPassword")
+const DBExplorer = ({
+  password,
+  tabIndex,
+  selectedOption
+}: DBExplorerProps) => {
 
-  useEffect(() => {
-
-    if (props.tabIndex !== 2) return
-    
-    window.databaseApi.receiveAllStakerAddresses(
-      (
-        event: Electron.IpcMainEvent,
-        result: number,
-        data: StakerMap,
-        errorMessage: string
-      ) => {
-        if (result === 0) {
-          
-          setAllStakers(data);
-          setAddressList(Object.keys(data));
-          setCurrAddress(Object.keys(data).length > 0 ? Object.keys(data)[0] : "")
-
-        } else {
-          console.error("Error AllStakerAddresses");
-          console.error(errorMessage);
-        }
-      }
-    );
-    window.databaseApi.reqAllStakerAddresses(dbPassword);
-
-  }, [props.tabIndex])
-
-  const currStaker = allStakers[currAddress]
-  const mnemonics = currStaker?.mnemonics ?? {}
-  const validators = currStaker?.validators ?? {}
-  const mnemonicCount = Object.keys(mnemonics).length;
-  const validatorCount = Object.keys(validators).length;
+  console.log("DBExplorer:", tabIndex, selectedOption)
 
   return (<>
-    <Center>
-      <Box sx={raisedWidgetStyle} bg="#2b2852">
-        <Box
-          width={'905px'}
-          height={'70vh'}
-          sx={{
-            border: '1px solid',
-            borderColor: 'purple.light',
-            padding: '24px',
-            borderRadius: '16px',
-          }}
-        >
-          { addressList.length === 0 && <Box><Text color={"white"}>There is no saved data in the database</Text></Box>}
-          { addressList.length > 0 && (
-            <>
-              <AddressSelect currAddress={currAddress} setCurrAddress={setCurrAddress} addressList={addressList} />
-              <Tabs
-                    index={selectedTab}
-                    onChange={(index) => setSelectedTab(index)}
-                  >
-                    <TabList>
-                      <Tab color={"white"}>Mnemonics</Tab>
-                      <Tab color={"white"}>Validators</Tab>
-                    </TabList>
-                    <TabPanels>
-                      <TabPanel sx={{ width: "100%" }}>
-                        <DataTable title="mnemonics" dataCount={mnemonicCount} data={mnemonics} /> 
-                      </TabPanel>
-                      <TabPanel sx={{ width: "100%" }}>
-                        <DataTable title="validators" dataCount={validatorCount} data={validators} />   
-                      </TabPanel>
-                    </TabPanels>
-              </Tabs>        
-            </>
-          )}
-        </Box>
-      </Box>
-    </Center></>
+    {/* <ScaleFade initialScale={0.5} in={tabIndex === 1}> */}
+      {/* {selectedOption === 0 && (
+        <ScaleFade initialScale={0.5} in={selectedOption === 0}>
+          <SavedDataWidget tabIndex={0} />
+        </ScaleFade>
+      )} */}
+      {selectedOption === 1 && (
+        // <ScaleFade initialScale={0.5} in={selectedOption === 1}>
+          <HistoryWidget tabIndex={1} />
+        // </ScaleFade>
+      )}
+     {/* </ScaleFade> */}
+    </>
   );
 };
